rkt: Split the stage1-image-related code to a separate file
This code is used not only by the run command, but also by prepare, so
let's move it outside. Also, it will grow in the next commit.
